services/presence/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-socketio==5.13.0
python-engineio==4.12.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
asyncpg==0.30.0
aio-pika==9.5.5
aiohttp==3.9.3
gunicorn==21.2.0
python-jose>=3.3.0
cryptography>=35.0.0